relocate: Ticket 의 위치를 이동
BREAKING CHANGE:
실질적인 데이터를 저장하므로 구별을 위해 model 의 하위 패키지로 이동합니다.
